rename blueprint group file to fix spelling error
